+++ exercicios/lista02/q5.c
@@ -4,23 +4,23 @@
 #define PI 3.1415
 
 // Funcao que calcula o quadrado de um numero qualquer
-double elevarQuadrado(double n){
-    double resultado;
+float elevarQuadrado(float n){
+    float resultado;
     resultado = n * n;
     return resultado;
 }
 
 // Função que calcula a area do circulo
-double areaCirculo(double raio){
-    double raioQuadrado = elevarQuadrado(raio);
-    double area = PI * raioQuadrado;
+float areaCirculo(float raio){
+    float raioQuadrado = elevarQuadrado(raio);
+    float area = PI * raioQuadrado;
     return area;
 }
 
 int main(){
-    double raio;
+    float raio;
     printf("Digite o raio do circulo em cm para calcular a area:\n");
-    scanf("%lf", &raio);
+    scanf("%f", &raio);
     printf("A area do circulo eh: %.2fcm^2", areaCirculo(raio));
     return 0;
 }
