+++ exercicios/lista02/q8.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 
 // Funcao que calcula faz o cambio para dolar e euro
-void cambio(double valor, double taxaDolar, double taxaEuro, double *cambioDolar, double *cambioEuro){
+void cambio(float valor, float taxaDolar, float taxaEuro, float *cambioDolar, float *cambioEuro){
     *cambioDolar = valor*taxaDolar;
     *cambioEuro = valor*taxaEuro;
 }
 
 int main(){
-    double valor, cotDolar, cotEuro, cambioDolar, cambioEuro; 
+    float valor, cotDolar, cotEuro, cambioDolar, cambioEuro; 
     printf("Digite o valor que voce quer converter\n");
-    scanf("%lf", &valor);
+    scanf("%f", &valor);
     printf("Digite a cotacao do dolar\n");
-    scanf("%lf", &cotDolar);
+    scanf("%f", &cotDolar);
     printf("Digite a cotacao do euro\n");
-    scanf("%lf", &cotEuro);
+    scanf("%f", &cotEuro);
     cambio(valor, cotDolar, cotEuro, &cambioDolar, &cambioEuro);
-    printf("O valor em dolares eh de: %.2lf, e em euros eh: %.2lf", cambioDolar, cambioEuro);
+    printf("O valor em dolares eh de: %.2f, e em euros eh: %.2f", cambioDolar, cambioEuro);
 }
